LadderGame 의 출력 관련 로직을 OutputView 로 이동
출력 관련 책임을 OutputView 에 위임하여 LadderGame 클래스가
게임 전체 흐름에만 집중할 수 있도록 변경하였습니다.
